Update lci-EP2050.csv
Changed area of panel per 3kWp from 370.8 m2/3kWp to 12.36 m2/3kWp. 
This is the results of changes in assumed lifetimes. For the 3kWp installation, a lifetime of 30 years is assumed. The required area of 370.8 m2 tandem/3kWp installation is obtained when assuming a 1 year lifetime for the tandem (i.e. requiring 30 tandem panels per 3kWp installation). The required area of 12.36 m2 tandem/3kWp installation is obtained when assuming a 30 year lifetime for the tandem. The lifetime of the tandem is very uncertain, but for comparison we are interested in a hypothetical case that the tandem panel has a lifetime equal to that of the incumbent.